ENH: Updates for better unit test configuration, standardizing the discovery tests and fixing some issues.
