+++ .coveragerc
@@ -1,0 +1,2 @@
+[run]
+omit =./act/tests/*, ./act/*version*py
